Refactor out all caching functionality to trait
